proj3/backend/requirements.txt
tensorflow==2.0.0
numpy==1.18.5
pillow==7.2.0
keras==2.3.1
flask==1.1.2
flask_cors==3.0.8
